Refactored the company classes to be in proper packages.
